Moved location stuff into base class
