+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/APT_miniProject_Android.iml" filepath="$PROJECT_DIR$/APT_miniProject_Android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/APT_miniProject_Android.iml" filepath="$PROJECT_DIR$/.idea/APT_miniProject_Android.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
